requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==1.5.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
html5lib==0.9999999
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
nbformat==4.4.0
numpy==1.14.5
opencv-python==4.1.0.25
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==3.10.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.22.0
retrying==1.3.3
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
style==1.1.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
torch==1.1.0
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
update==0.0.1
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
